+++ tensorcircuit/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 __author__ = "TensorCircuit Authors"
 __creator__ = "refraction-ray"